[iOS] Shift debounce Obj-C wrappers into debounce component
This change moves the Obj-C wrappers for `DebounceService` into the debounce component in an `ios` layer.
