+++ components/debounce/ios/browser/BUILD.gn
@@ -1,13 +1,13 @@
-# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# Copyright (c) 2025 The Brave Authors. All rights reserved.
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
 
-source_set("debounce") {
+source_set("browser") {
   sources = [
-    "debounce_service+private.h",
-    "debounce_service.h",
-    "debounce_service.mm",
+    "debounce_service_bridge.h",
+    "debounce_service_bridge_impl.h",
+    "debounce_service_bridge_impl.mm",
   ]
   deps = [
     "//base",
