+++ components/debounce/ios/browser/headers.gni
@@ -1,7 +1,7 @@
-# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# Copyright (c) 2025 The Brave Authors. All rights reserved.
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
 
-browser_api_debounce_public_headers =
-    [ "//brave/ios/browser/api/debounce/debounce_service.h" ]
+debounce_component_public_headers =
+    [ "//brave/components/debounce/ios/browser/debounce_service_bridge.h" ]
