+++ ios/BUILD.gn
@@ -8,6 +8,7 @@ import("//brave/build/ios/config.gni")
 import("//brave/build/ios/create_xcframework.gni")
 import("//brave/build/mac/tweak_info_plist.gni")
 import("//brave/components/ai_chat/ios/browser/headers.gni")
+import("//brave/components/debounce/ios/browser/headers.gni")
 import("//brave/ios/browser/api/ads/headers.gni")
 import("//brave/ios/browser/api/brave_domains/headers.gni")
 import("//brave/ios/browser/api/brave_rewards/headers.gni")
@@ -20,7 +21,6 @@ import("//brave/ios/browser/api/certificate/headers.gni")
 import("//brave/ios/browser/api/content_settings/headers.gni")
 import("//brave/ios/browser/api/credential_provider/headers.gni")
 import("//brave/ios/browser/api/de_amp/headers.gni")
-import("//brave/ios/browser/api/debounce/headers.gni")
 import("//brave/ios/browser/api/developer_options_code/headers.gni")
 import("//brave/ios/browser/api/favicon/headers.gni")
 import("//brave/ios/browser/api/features/headers.gni")
@@ -104,7 +104,7 @@ brave_core_public_headers += brave_shields_public_headers
 brave_core_public_headers += browser_api_url_sanitizer_public_headers
 brave_core_public_headers += browser_url_sanitizer_public_headers
 brave_core_public_headers += browser_api_de_amp_public_headers
-brave_core_public_headers += browser_api_debounce_public_headers
+brave_core_public_headers += debounce_component_public_headers
 brave_core_public_headers +=
     browser_api_https_upgrade_exceptions_service_public_headers
 brave_core_public_headers +=
