+++ ios/brave-ios/Sources/Brave/Frontend/Settings/Features/ShieldsPrivacy/AdvancedShieldSettings.swift
@@ -106,7 +106,7 @@ import os
   private var subscriptions: [AnyCancellable] = []
   private let p3aUtilities: BraveP3AUtils
   private let deAmpPrefs: DeAmpPrefs
-  private let debounceService: DebounceService?
+  private let debounceService: (any DebounceService)?
   private let rewards: BraveRewards?
   private let clearDataCallback: ClearDataCallback
   private let webcompatReporterHandler: WebcompatReporterWebcompatReporterHandler?
@@ -116,7 +116,7 @@ import os
     profile: LegacyBrowserProfile,
     tabManager: TabManager,
     feedDataSource: FeedDataSource,
-    debounceService: DebounceService?,
+    debounceService: (any DebounceService)?,
     braveCore: BraveProfileController,
     p3aUtils: BraveP3AUtils,
     rewards: BraveRewards?,
