--- ios/browser/api/debounce/DEPS
@@ -1,3 +1,0 @@
-include_rules = [
-  "+brave/components/debounce/core/browser",
-]
