+++ ios/browser/debounce/DEPS
@@ -1,4 +1,5 @@
 include_rules = [
   "+brave/components/debounce/core/browser",
   "+brave/components/debounce/core/common",
+  "+brave/components/debounce/ios/browser",
 ]
